Splitting up responsibilities - making it more testable also
